Move `soft_dotted` check from Universal to Shaping profile
