+++ tests/profiles/shaping_test.py
@@ -5,6 +5,7 @@ import tempfile
 
 from fontbakery.checkrunner import FAIL, WARN
 from fontbakery.codetesting import (assert_PASS,
+                                    assert_SKIP,
                                     assert_results_contain,
                                     CheckTester,
                                     TEST_FILE)
@@ -116,3 +117,28 @@ def test_check_dotted_circle():
     font = TEST_FILE("broken_markazitext/MarkaziText-VF.ttf")
     assert_results_contain(check(font),
                            FAIL, "unattached-dotted-circle-marks")
+
+
+def test_check_soft_dotted():
+    """Check if font substitues soft dotted glyphs
+    when combined with top marks."""
+    check = CheckTester(shaping_profile,
+                        "com.google.fonts/check/soft_dotted")
+
+    font = TEST_FILE("abeezee/ABeeZee-Regular.ttf")
+    msg = assert_results_contain(check(font), WARN, "soft-dotted")
+    assert "The dot of soft dotted characters used in orthographies" not in msg
+    assert "The dot of soft dotted characters should disappear" in msg
+
+    font = TEST_FILE("cabin/Cabin-Regular.ttf")
+    msg = assert_results_contain(check(font), FAIL, "soft-dotted")
+    assert "The dot of soft dotted characters used in orthographies" in msg
+    assert "The dot of soft dotted characters should disappear" in msg
+
+    font = TEST_FILE("akshar/Akshar[wght].ttf")
+    assert_PASS(check(font),
+                "All soft dotted characters seem to lose their dot ...")
+
+    font = TEST_FILE("rosarivo/Rosarivo-Regular.ttf")
+    assert_SKIP(check(font),
+                "It is not clear if soft dotted characters ...")
